Add support for Python 3.7
There is actually no syntactic difference between 3.6 and 3.7,
*except* that in 3.7 'async' and 'await' are always keywords.  So this
diff simply adds a check that forces these two to be
keywords when feature_version >= 7.

There is one place where the generated AST is different between 3.6
and 3.7: where 3.6 has

  comp_for: [ASYNC] 'for' exprlist 'in' or_test [comp_iter]

3.7 has

  sync_comp_for: 'for' exprlist 'in' or_test [comp_iter]
  comp_for: ['async'] sync_comp_for

However, changing the AST structure to conform to 3.7 would be a
backwards incompatible change to typed_ast, while the current change
is not backwards incompatible (it simply adds a new flag bit
definition).

Finally, a note for mypy specifically: when this version in installed,
using Python 3.7 with mypy will reject code that was previously (and
wrongly) not rejected: in particular code that uses 'async' or 'await'
as a non-keyword is invalid in Python 3.7, but mypy would not have
rejected that code with previous typed_ast versions.  Does this mean
we should bump the version to 1.3?  I don't know...