Fix f-string parsing to not think everything is a set
typed-ast has (for a while) done f-string weirdly, essentially parsing
the f string contents twice. This was dodgy but harmless until a 3.7
change tweaked the string that was being used for the second parse to
be enclosed by braces, which is causing typed_ast to think that every
value in an f-string is a set.

Restore things to working like cpython.